src/config/requirements.txt
# Enterprise Guardrails and Compliance Requirements
# Core dependencies for guardrails, security, and compliance functionality

# Encryption and Security
cryptography>=41.0.0
PyJWT>=2.8.0

# Data Processing and Analysis
pydantic>=2.0.0
python-dateutil>=2.8.0

# PII Detection (would integrate with spaCy or transformers for NER)
# spacy>=3.6.0
# transformers>=4.30.0

# Configuration Management
PyYAML>=6.0

# Logging and Monitoring
structlog>=23.0.0

# Testing (for development)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Type Checking
mypy>=1.5.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0